# perf: cache layout access check