Specifying a layout on render causes a synchronous `fs.accessSync` on every request. Cache the result of these lookups in LRU to avoid duplicate blocking checks.

### Benchmarks (master)
```
fastify-ejs-local-layout.js: 84885.34 req/s
fastify-ejs-global-layout.js: 115680 req/s
```

### Benchmarks (PR)
```
fastify-ejs-local-layout.js: 114592 req/s
fastify-ejs-global-layout.js: 116917.34 req/s
```

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)