@gurgunday 2024-02-27
Can you replace hashlru with toad-cache (_LruMap_) in another PR?

We try to use that for all the packages now, unless there is a significant perf regression