@mweberxyz 2024-02-28
> Actually, just checked the PR again and it looks to me like this function gets called once?
> 
> Is that not the case?

It gets called a single time for a globally defined layout.

It gets called on every request (via `layoutIsValid`) for a reply-defined layout, ie -
```
reply.view('index-for-layout.ejs', { text: 'text' }, { layout: 'layout.html' })
```