@mweberxyz 2024-02-28
Marked as draft to let https://github.com/fastify/point-of-view/pull/416 and the subsequent `toad-lru` PR merge first to avoid conflict.